+++ src/main/java/com/mockproject/service/SyllabusService.java
@@ -3,6 +3,7 @@ package com.mockproject.service;
 import com.mockproject.dto.SessionDTO;
 import com.mockproject.dto.SyllabusDTO;
 import com.mockproject.entity.*;
+import com.mockproject.mapper.SessionMapper;
 import com.mockproject.mapper.SyllabusMapper;
 import com.mockproject.repository.OutputStandardRepository;
 import com.mockproject.repository.SyllabusRepository;
@@ -253,8 +254,16 @@ public class SyllabusService implements ISyllabusService {
     public boolean deleteSyllabus(Long syllabusId, boolean status){
         Optional<Syllabus> syllabus = syllabusRepository.findByIdAndStatus(syllabusId, status);
         syllabus.orElseThrow(() -> new ResponseStatusException(HttpStatus.NO_CONTENT,"Syllabus not found"));
+        List<SessionDTO> sessionDTOList = sessionService.getAllSessionBySyllabusId(syllabusId, true);
+        List<Session>   sessionList = new ArrayList<>();
+
+        for(SessionDTO sessionDTO : sessionDTOList){
+            sessionList.add(SessionMapper.INSTANCE.toEntity(sessionDTO));
+        }
+        syllabus.get().setListSessions(sessionList);
         syllabus.get().setStatus(false);
-        sessionService.deleteSessions(syllabusId, status);
+        if(!(syllabus.get().getListSessions().isEmpty()))
+            sessionService.deleteSessions(syllabusId, status);
         syllabusRepository.save(syllabus.get());
         return true;
     }
