+++ src/main/java/com/mockproject/service/UnitDetailService.java
@@ -1,13 +1,14 @@
 package com.mockproject.service;
 
+import com.mockproject.dto.SessionDTO;
 import com.mockproject.dto.TrainingMaterialDTO;
 import com.mockproject.dto.UnitDTO;
 import com.mockproject.dto.UnitDetailDTO;
-import com.mockproject.entity.CustomUserDetails;
-import com.mockproject.entity.Unit;
-import com.mockproject.entity.UnitDetail;
-import com.mockproject.entity.User;
+import com.mockproject.entity.*;
+import com.mockproject.mapper.SessionMapper;
+import com.mockproject.mapper.TrainingMaterialMapper;
 import com.mockproject.mapper.UnitDetailMapper;
+import com.mockproject.repository.SyllabusRepository;
 import com.mockproject.repository.UnitDetailRepository;
 import com.mockproject.repository.UnitRepository;
 import com.mockproject.service.interfaces.ITrainingMaterialService;
@@ -38,6 +39,8 @@ public class UnitDetailService implements IUnitDetailService {
     private final UnitDetailRepository unitDetailRepository;
 
     private final ITrainingMaterialService trainingMaterialService;
+
+    private final SyllabusRepository syllabusRepository;
 
     @Override
     public List<UnitDetailDTO> getUnitDetailByUnitId(Long idUnit) {
@@ -129,8 +132,24 @@ public class UnitDetailService implements IUnitDetailService {
     public boolean deleteUnitDetail(Long unitDetailId, boolean status){
         Optional<UnitDetail> unitDetail = unitDetailRepository.findByIdAndStatus(unitDetailId, status);
         unitDetail.orElseThrow(() -> new ResponseStatusException(HttpStatus.NO_CONTENT, "UnitDetail not found"));
+        List<TrainingMaterialDTO> trainingMaterialDTOList = trainingMaterialService.getFiles(unitDetailId, true);
+        List<TrainingMaterial>   trainingMaterialList = new ArrayList<>();
+
+        for(TrainingMaterialDTO trainingMaterialDTO : trainingMaterialDTOList){
+            trainingMaterialList.add(TrainingMaterialMapper.INSTANCE.toEntity(trainingMaterialDTO));
+        }
+        unitDetail.get().setListMaterials(trainingMaterialList);
         unitDetail.get().setStatus(false);
-        trainingMaterialService.deleteTrainingMaterials(unitDetailId,status);
+        if(!unitDetail.get().getListMaterials().isEmpty())
+            trainingMaterialService.deleteTrainingMaterials(unitDetailId,status);
+
+        unitDetail.get().getUnit().setDuration(unitDetail.get().getUnit().getDuration().subtract(unitDetail.get().getDuration().divide(BigDecimal.valueOf(60))));
+
+        unitDetail.get().getUnit().getSession().getSyllabus().setHour(unitDetail.get().getUnit().getSession().getSyllabus().getHour().subtract(unitDetail.get().getDuration().divide(BigDecimal.valueOf(60))));
+
+        unitRepository.save(unitDetail.get().getUnit());
+        syllabusRepository.save(unitDetail.get().getUnit().getSession().getSyllabus());
+
         unitDetailRepository.save(unitDetail.get());
         return true;
     }
@@ -138,6 +157,7 @@ public class UnitDetailService implements IUnitDetailService {
     @Override
     public boolean deleteUnitDetails(Long unitId, boolean status){
         Optional<List<UnitDetail>> unitDetails = unitDetailRepository.findByUnitIdAndStatus(unitId, status);
+        System.out.println("Dang o Unitdetail");
         ListUtils.checkList(unitDetails);
         unitDetails.get().forEach((i) -> deleteUnitDetail(i.getId(), status));
         return true;
